Adds a function to recursively apply functors to datasets
Pytorch datasets have no more interface than a list. However
during splitting they often become hidden behind a hierarchy of
Subset proxy objects and ConcatDataset composite objects.
This makes e.g. adding a transform after constructing the inital
dataset cumbersome. This is unfortunate as
1) Creating 2) Splitting 3) Computing Normalization on Trainset
is a usual workflow. To enable this a new function is added to
recursively traverse the tree of datasets.